Colour change in the chord diagram for accessibility
To meet the accessibility guideline, some changes in colours used in the
chord diagram for `Interprovincial mobility origin and destination` was
recommended.

To avoid having green and red together for colour blindness, Ontario's
colour is changed to orange (#D55E00).
For a higher contrast, Territories' colour is changed to
darker(#926AA6).