Revise data via GitHub Actions
Using GitHub Actions, downloading and processing data after releasing
the table is now semi-automated.
After the new data being released on the public website, initiating the
process through Actions will revise the data in this repository. This
can be useful to share the workload to others who are not familiar with
running R and using git.